feat: extract initialize logic
