+++ packages/astro/src/assets/fonts/types.ts
@@ -44,3 +44,17 @@ export type GoogleProviderName = typeof GOOGLE_PROVIDER_NAME;
 export type BuiltInProvider = (typeof BUILTIN_PROVIDERS)[number];
 
 export type FontType = (typeof FONT_TYPES)[number];
+
+/**
+ * Preload data is used for links generation inside the <Font /> component
+ */
+export type PreloadData = Array<{
+	/**
+	 * Absolute link to a font file, eg. /_astro/fonts/abc.woff
+	 */
+	url: string;
+	/**
+	 * A font type, eg. woff2, woff, ttf...
+	 */
+	type: FontType;
+}>;
