+++ packages/astro/src/assets/fonts/utils.ts
@@ -4,6 +4,7 @@ import { extname } from 'node:path';
 import { DEFAULT_FALLBACKS, FONT_TYPES } from './constants.js';
 import type { Storage } from 'unstorage';
 import type * as fontaine from 'fontaine';
+import type { Logger } from '../../core/logger/core.js';
 
 // TODO: expose all relevant options in config
 // Source: https://github.com/nuxt/fonts/blob/main/src/css/render.ts#L7-L21
@@ -176,9 +177,59 @@ export async function generateFallbacksCSS({
 			font: fallback,
 			// TODO: support family.as
 			name: `${family} fallback: ${fallback}`,
-			metrics: (await getMetricsForFamily(fallback, null)) ?? undefined,
 		});
 	}
 
 	return { css, fallbacks };
+}
+
+/**
+ * We want to show logs related to font downloading (fresh or from cache)
+ * However if we just use the logger as is, there are too many logs, and not
+ * so useful.
+ * This log manager allows avoiding repetitive logs:
+ * - If there are many downloads started at once, only one log is shown for start and end
+ * - If a given file has already been logged, it won't show up anymore (useful in dev)
+ */
+// TODO: test
+export function createLogManager(logger: Logger) {
+	const done = new Set<string>();
+	const items = new Set<string>();
+	let id: NodeJS.Timeout | null = null;
+
+	return {
+		add: (value: string) => {
+			if (done.has(value)) {
+				return;
+			}
+
+			if (items.size === 0 && id === null) {
+				logger.info('assets', 'Downloading fonts...');
+			}
+			items.add(value);
+			if (id) {
+				clearTimeout(id);
+				id = null;
+			}
+		},
+		remove: (value: string, cached: boolean) => {
+			if (done.has(value)) {
+				return;
+			}
+
+			items.delete(value);
+			done.add(value);
+			if (id) {
+				clearTimeout(id);
+				id = null;
+			}
+			id = setTimeout(() => {
+				let msg = 'Done';
+				if (cached) {
+					msg += ' (loaded from cache)';
+				}
+				logger.info('assets', msg);
+			}, 50);
+		},
+	};
 }
